meson: Sync Nasm fallback with libvpx
Part-of: <https://gitlab.freedesktop.org/gstreamer/meson-ports/ffmpeg/-/merge_requests/95>

+++ subprojects/win-nasm/download-binary.py
@@ -1,4 +1,7 @@
 #!/usr/bin/env python3
+# SPDX-FileCopyrightText: 2020 Nirbheek Chauhan <nirbheek@centricular.com>
+# SPDX-FileCopyrightText: 2023 L. E. Segovia <amy@centricular.com>
+# SPDX-License-Ref: LGPL-2.1-or-later
 
 import os
 import sys
@@ -19,7 +22,12 @@ UPSTREAM_URL = 'https://www.nasm.us/pub/nasm/releasebuilds/{}/{}/{}'
 GSTREAMER_URL = 'https://gstreamer.freedesktop.org/src/mirror/{}'
 
 version = sys.argv[1]
-arch = 'win64' if sys.argv[2] == 'x86_64' else 'win32'
+if sys.argv[2] == 'darwin':
+    arch = 'macosx'
+elif sys.argv[2] == 'x86_64':
+    arch = 'win64'
+else:
+    arch = 'win32'
 zip_sha256 = sys.argv[3]
 source_dir = os.path.join(os.environ['MESON_SOURCE_ROOT'], os.environ['MESON_SUBDIR'])
 dest = BASENAME.format(version, arch)
@@ -35,30 +43,32 @@ if os.path.isfile(dest_path):
     found_sha256 = get_sha256(dest_path)
     if found_sha256 == zip_sha256:
         print('{} already downloaded'.format(dest))
-        sys.exit(0)
     else:
-        print('{} checksum mismatch, redownloading'.format(dest))
+        os.remove(dest)
+        print('{} checksum mismatch, redownloading'.format(dest), file=sys.stderr)
 
-for url in (GSTREAMER_URL.format(dest), UPSTREAM_URL.format(version, arch, dest)):
-    print('Downloading {} to {}'.format(url, dest))
-    try:
-        with open(dest_path, 'wb') as d:
-            f = urllib.request.urlopen(url, context=ctx)
-            d.write(f.read())
-        break
-    except urllib.error.URLError as ex:
-        print(ex)
-        print('Failed to download from {!r}, trying mirror...'.format(url))
-        continue
-else:
-    curdir = os.path.dirname(sys.argv[0])
-    print('Couldn\'t download {!r}! Try downloading it manually and '
-          'placing it into {!r}'.format(dest, curdir))
+if not os.path.isfile(dest_path):
+    for url in (GSTREAMER_URL.format(dest), UPSTREAM_URL.format(version, arch, dest)):
+        print('Downloading {} to {}'.format(url, dest))
+        try:
+            with open(dest_path, 'wb') as d:
+                f = urllib.request.urlopen(url, context=ctx)
+                d.write(f.read())
+            break
+        except BaseException as ex:
+            print(ex, file=sys.stderr)
+            print('Failed to download from {!r}, trying mirror...'.format(url), file=sys.stderr)
+            continue
+    else:
+        curdir = os.path.dirname(sys.argv[0])
+        print('Couldn\'t download {!r}! Try downloading it manually and '
+            'placing it into {!r}'.format(dest, curdir), file=sys.stderr)
+        sys.exit(1)
 
 found_sha256 = get_sha256(dest_path)
 if found_sha256 != zip_sha256:
     print('SHA256 of downloaded file {} was {} instead of {}'
-          ''.format(dest, found_sha256, zip_sha256))
+          ''.format(dest, found_sha256, zip_sha256), file=sys.stderr)
     sys.exit(1)
 
 print('Extracting {}'.format(dest))
